Replace RemoteStorage with a local git repo.
Some revision data is also stored in a database table - that lets us track
visibility-related metadata so we can eventually apply permissions checks
to individual notebooks.  Note that all revision and clickstream objects
get deleted when the notebook is deleted, so only git will have history of
deleted notebooks.

This fixes #26 but is only the beginning of #25.