[config, db] Remove dependency on the database during startup (#894)
* [config] Remove dependency on the database during startup

Attempting to load the application used to require a valid connection to
the database, which prevented the use of commands like `rails db:create`.

This was caused by a statement in the `User` model class that attempted
to connect to the database to find out the names of the fields in the
`UserSummary` class in order to define delegator methods for each of them.
A similar statement in the `Notebook` class had the same effect.

This change fixes the issue by allowing this delegation to happen later
during the application's runtime.

In the Notebook class, the delegated methods have been hard-coded instead
of being dynamically looked up. This is because some of them used a prefix
while others didn't. Thus, new attributes will note be delegated by default.

The User class was easier because it treated all delegated methods the same,
so it still allows for dynamic lookup. It also ensures that any methods defined
on the User class are not delegated. As a side effect, though, it may delegate
methods that are not attributes, but this seems like a small price to pay.

With this change, we can now run `rails db:create` and `db:test:prepare`.
We can also run `rails console` even if we don't have access to the
database.

* [db] Add the database schema to source control

It is strongly recommended to track database schema changes, showing
how these changes relate to the history of the project.

With the schema in the repository, we can now run `rails db:setup`,
`rails db:schema:load`, and `rails db:test:prepare` out of the box.